+++ changelog.md
@@ -1,3 +1,12 @@
+=================================
+- Fix Tasks::setFilters
+
+1.6.0
+=================================
+- Move Project query into context
+- Move Task query into context
+- Move Timer query into context
+
 1.5.6
 =================================
 - Added links